+++ src/alignment.rs
@@ -1,3 +1,14 @@
+//! This module provides functionality for determining optimal memory alignment
+//! for allocations based on the number of bytes requested. It includes constants
+//! for common alignment sizes and a struct to represent alignment hints.
+//!
+//! The main functionality is provided by the [`AlignmentHint::new`](AlignmentHint::get) 
+//! function, which returns an [`AlignmentHint`] indicating the suggested alignment and whether
+//! huge pages should be used.
+//!
+//! # Structs
+//! - [`AlignmentHint`]: Contains alignment information and a flag for huge page usage.
+
 const TWO_MEGABYTES: usize = 2 * 1024 * 1024;
 const SIXTY_FOUR_BYTES: usize = 64;
 const ZERO_BYTES: usize = 0;
@@ -14,34 +25,37 @@ pub struct AlignmentHint {
     pub use_huge_pages: bool,
 }
 
-/// Gets the optimal alignment for the number of bytes.
-///
-/// If the number of bytes is a multiple of 2 MB, a natural 2 MB boundary
-/// is selected and a hint for using Huge/Large Pages is issued.
-///
-/// In any other case, an alignment of 64 byte boundaries is produced, which
-/// should be optimal for both AVX-2 and AVX-512 operations.
-///
-/// ## Arguments
-/// * `num_bytes` - The number of bytes to allocate.
-///
-/// ## Returns
-/// An [`AlignmentHint`] outlining the suggested alignment.
-pub fn get_alignment(num_bytes: usize) -> AlignmentHint {
-    if num_bytes == 0 {
-        AlignmentHint {
-            alignment: ZERO_BYTES,
-            use_huge_pages: false,
-        }
-    } else if (num_bytes & (TWO_MEGABYTES - 1)) == 0 {
-        AlignmentHint {
-            alignment: TWO_MEGABYTES,
-            use_huge_pages: true,
-        }
-    } else {
-        AlignmentHint {
-            alignment: SIXTY_FOUR_BYTES,
-            use_huge_pages: false,
+impl AlignmentHint {
+    /// Gets the optimal alignment for the number of bytes.
+    ///
+    /// If the number of bytes is a multiple of 2 MB, a natural 2 MB boundary
+    /// is selected and a hint for using Huge/Large Pages is issued.
+    ///
+    /// In any other case, an alignment of 64 byte boundaries is produced, which
+    /// should be optimal for both AVX-2 and AVX-512 operations.
+    ///
+    /// ## Arguments
+    /// * `num_bytes` - The number of bytes to allocate.
+    ///
+    /// ## Returns
+    /// An [`AlignmentHint`] outlining the suggested alignment.
+    #[inline(always)]
+    pub fn new(num_bytes: usize) -> Self {
+        if num_bytes == 0 {
+            AlignmentHint {
+                alignment: ZERO_BYTES,
+                use_huge_pages: false,
+            }
+        } else if (num_bytes & (TWO_MEGABYTES - 1)) == 0 {
+            AlignmentHint {
+                alignment: TWO_MEGABYTES,
+                use_huge_pages: true,
+            }
+        } else {
+            AlignmentHint {
+                alignment: SIXTY_FOUR_BYTES,
+                use_huge_pages: false,
+            }
         }
     }
 }
@@ -52,42 +66,42 @@ mod tests {
 
     #[test]
     fn get_alignment_4mb_is_2mb_aligned_hugepage() {
-        let memory = get_alignment(2 * TWO_MEGABYTES);
+        let memory = AlignmentHint::new(2 * TWO_MEGABYTES);
         assert_eq!(memory.alignment, TWO_MEGABYTES);
         assert!(memory.use_huge_pages);
     }
 
     #[test]
     fn get_alignment_2mb_is_2mb_aligned_hugepage() {
-        let memory = get_alignment(TWO_MEGABYTES);
+        let memory = AlignmentHint::new(TWO_MEGABYTES);
         assert_eq!(memory.alignment, TWO_MEGABYTES);
         assert!(memory.use_huge_pages);
     }
 
     #[test]
     fn get_alignment_1mb_is_64b_aligned() {
-        let memory = get_alignment(TWO_MEGABYTES / 2);
+        let memory = AlignmentHint::new(TWO_MEGABYTES / 2);
         assert_eq!(memory.alignment, SIXTY_FOUR_BYTES);
         assert!(!memory.use_huge_pages);
     }
 
     #[test]
     fn get_alignment_63kb_is_64b_aligned() {
-        let memory = get_alignment(63 * 1024);
+        let memory = AlignmentHint::new(63 * 1024);
         assert_eq!(memory.alignment, SIXTY_FOUR_BYTES);
         assert!(!memory.use_huge_pages);
     }
 
     #[test]
     fn get_alignment_64kb_is_64b_aligned() {
-        let memory = get_alignment(64 * 1024);
+        let memory = AlignmentHint::new(64 * 1024);
         assert_eq!(memory.alignment, SIXTY_FOUR_BYTES);
         assert!(!memory.use_huge_pages);
     }
 
     #[test]
     fn get_alignment_0b_is_0b_aligned() {
-        let memory = get_alignment(0);
+        let memory = AlignmentHint::new(0);
         assert_eq!(memory.alignment, ZERO_BYTES);
         assert!(!memory.use_huge_pages);
     }